% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{count_foci}
\alias{count_foci}
\title{Count foci}
\usage{
count_foci(dir = NA, sample_cells = F, dose = "ugml", time = "min")
}
\arguments{
\item{dir}{ImageJ output directory containing uncounted csv files.}

\item{sample_cells}{To perform downsampling on cells to N number of cells. Default is FALSE.}

\item{dose}{The unit of stimulation concentration.
This must be written in the file name between underscores (e.g. "x_10ugml_x").}

\item{time}{The unit of time.
This must be written in the file name between underscores (e.g. "x_10min_x").}
}
\value{
data frame containing cell identifier and foci count.
}
\description{
Count foci
}
\examples{
count_foci("imagej_output")
}
