% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatscatter}
\alias{heatscatter}
\title{Shiny-based visualization of single-cell foci number across time}
\usage{
heatscatter()
}
\value{

}
\description{
Shiny-based visualization of single-cell foci number across time
}
\examples{
heatscatter()
}
